model_training/GeneralizedLoss-Counting-Pytorch-main/requirements.txt
line-profiler==3.3.0
Markdown==3.3.4
matplotlib
matplotlib-inline==0.1.2
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.2.54
opt-einsum==3.3.0
pafy==0.5.5
pandas==1.2.3
parso==0.8.2
pickleshare==0.7.5
Pillow
scipy==1.6.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
torch==1.9.0
torchvision==0.10.0
tornado
tqdm